Création du fichier et de la classe Categorie
